Support TypeConfiguration in provider-definition-schema and split the provider-definition-schema to base-schema to extend for other types
