+++ src/main/java/software/amazon/cloudformation/resource/Validator.java
@@ -15,57 +15,35 @@
 package software.amazon.cloudformation.resource;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 
 import lombok.Builder;
 
 import org.everit.json.schema.Schema;
 import org.everit.json.schema.loader.SchemaClient;
-import org.everit.json.schema.loader.SchemaLoader;
 import org.everit.json.schema.loader.SchemaLoader.SchemaLoaderBuilder;
 import org.everit.json.schema.loader.internal.DefaultSchemaClient;
-import org.json.JSONException;
 import org.json.JSONObject;
-import org.json.JSONTokener;
 
 import software.amazon.cloudformation.resource.exceptions.ValidationException;
 
-public class Validator implements SchemaValidator {
-    private static final URI JSON_SCHEMA_URI_HTTP = newURI("http://json-schema.org/draft-07/schema");
+public class Validator extends BaseValidator {
+
+    protected static final String RESOURCE_DEFINITION_SCHEMA_PATH = "/schema/"
+        + "provider.definition.schema.v1.json";
+    protected static final String TYPE_CONFIGURATION_DEFINITION_SCHEMA_PATH = "/schema/"
+        + "provider.configuration.definition.schema.v1.json";
     private static final URI RESOURCE_DEFINITION_SCHEMA_URI = newURI(
         "https://schema.cloudformation.us-east-1.amazonaws.com/provider.definition.schema.v1.json");
-    private static final String ID_KEY = "$id";
-    private static final String JSON_SCHEMA_PATH = "/schema/schema";
-    private static final String RESOURCE_DEFINITION_SCHEMA_PATH = "/schema/provider.definition.schema.v1.json";
-
-    /**
-     * resource definition schema ("resource schema schema"). All resource schemas
-     * are validated against this one and JSON schema draft v7 below.
-     */
-    private final JSONObject definitionSchemaJsonObject;
-
-    /**
-     * locally cached draft-07 JSON schema. All resource schemas are validated
-     * against it
-     */
-    private final JSONObject jsonSchemaObject;
-    /**
-     * this is what SchemaLoader uses to download remote $refs. Not necessarily an
-     * HTTP client, see the docs for details. We override the default SchemaClient
-     * client in unit tests to be able to control how remote refs are resolved.
-     */
-    private final SchemaClient downloader;
+    private JSONObject typeConfigurationDefinitionJson;
 
     public Validator(SchemaClient downloader) {
-        this(loadResourceAsJSON(JSON_SCHEMA_PATH), loadResourceAsJSON(RESOURCE_DEFINITION_SCHEMA_PATH), downloader);
+        this(loadResourceAsJSON(RESOURCE_DEFINITION_SCHEMA_PATH), downloader);
+        this.typeConfigurationDefinitionJson = loadResourceAsJSON(TYPE_CONFIGURATION_DEFINITION_SCHEMA_PATH);
     }
 
-    private Validator(JSONObject jsonSchema,
-                      JSONObject definitionSchema,
+    private Validator(JSONObject definitionSchema,
                       SchemaClient downloader) {
-        this.jsonSchemaObject = jsonSchema;
-        this.definitionSchemaJsonObject = definitionSchema;
-        this.downloader = downloader;
+        super(definitionSchema, downloader);
     }
 
     @Builder
@@ -78,25 +56,13 @@ public class Validator implements SchemaValidator {
      */
     private Schema makeResourceDefinitionSchema() {
         SchemaLoaderBuilder builder = getSchemaLoader();
+        registerMetaSchema(builder, typeConfigurationDefinitionJson);
         builder.schemaJson(definitionSchemaJsonObject);
         return builder.build().load().build();
     }
 
-    @Override
-    public void validateObject(final JSONObject modelObject, final JSONObject definitionSchemaObject) throws ValidationException {
-        final SchemaLoaderBuilder loader = getSchemaLoader(definitionSchemaObject);
-
-        try {
-            final Schema schema = loader.build().load().build();
-            schema.validate(modelObject); // throws a ValidationException if this object is invalid
-        } catch (final org.everit.json.schema.ValidationException e) {
-            throw ValidationException.newScrubbedException(e);
-        }
-    }
-
     /**
-     * Performs JSON Schema validation for the input resource definition against the
-     * resource provider definition schema
+     * Performs JSON Schema validation for the input resource definition against the resource provider definition schema
      *
      * @param resourceDefinition JSON-encoded resource definition
      * @throws ValidationException Thrown for any schema validation errors
@@ -110,8 +76,8 @@ public class Validator implements SchemaValidator {
      * create a Schema instance that can be used to validate CloudFormation resources.
      *
      * @param resourceDefinition - CloudFormation Resource Provider Schema (Resource Definition)
-     * @throws ValidationException if supplied <code>resourceDefinition</code> is invalid.
      * @return - Schema instance for the given Resource Definition
+     * @throws ValidationException if supplied <code>resourceDefinition</code> is invalid.
      */
     public Schema loadResourceDefinitionSchema(final JSONObject resourceDefinition) {
 
@@ -131,7 +97,8 @@ public class Validator implements SchemaValidator {
             // step 2: load resource definition as a Schema that can be used to validate resource models;
             // definitionSchemaJsonObject becomes a meta-schema
             SchemaLoaderBuilder builder = getSchemaLoader();
-            registerMetaSchema(builder, jsonSchemaObject);
+            registerMetaSchema(builder, resourceDefinition);
+            registerMetaSchema(builder, typeConfigurationDefinitionJson);
             builder.schemaJson(resourceDefinition);
             // when resource definition is loaded as a schema, $refs are resolved and validated
             return builder.build().load().build();
@@ -139,68 +106,6 @@ public class Validator implements SchemaValidator {
             throw ValidationException.newScrubbedException(e);
         } catch (final org.everit.json.schema.SchemaException e) {
             throw new ValidationException(e.getMessage(), e.getSchemaLocation(), e);
-        }
-    }
-
-    /**
-     * Convenience method - creates a SchemaLoaderBuilder with cached JSON draft-07 meta-schema
-     *
-     * @param schemaObject
-     * @return
-     */
-    private SchemaLoaderBuilder getSchemaLoader(JSONObject schemaObject) {
-        return getSchemaLoader().schemaJson(schemaObject);
-    }
-
-    /** get schema-builder preloaded with JSON draft V7 meta-schema */
-    private SchemaLoaderBuilder getSchemaLoader() {
-        final SchemaLoaderBuilder builder = SchemaLoader
-            .builder()
-            .draftV7Support()
-            .schemaClient(downloader);
-
-        // registers the local schema with the draft-07 url
-        builder.registerSchemaByURI(JSON_SCHEMA_URI_HTTP, jsonSchemaObject);
-
-        return builder;
-    }
-
-    private static JSONObject loadResourceAsJSON(String path) {
-        return new JSONObject(new JSONTokener(Validator.class.getResourceAsStream(path)));
-    }
-
-    /** wrapper around new URI that throws an unchecked exception */
-    static URI newURI(final String uri) {
-        try {
-            return new URI(uri);
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(uri);
-        }
-    }
-
-    /**
-     * Register a meta-schema with the SchemaLoaderBuilder. The meta-schema $id is used to generate schema URI
-     * This has the effect of caching the meta-schema. When SchemaLoaderBuilder is used to build the Schema object,
-     * the cached version will be used. No calls to remote URLs will be made.
-     * Validator caches JSON schema (/resources/schema) and Resource Definition Schema
-     * (/resources/provider.definition.schema.v1.json)
-     *
-     * @param loaderBuilder
-     * @param schema meta-schema JSONObject to be cached. Must have a valid $id property
-     */
-    void registerMetaSchema(final SchemaLoaderBuilder loaderBuilder, JSONObject schema) {
-        try {
-            String id = schema.getString(ID_KEY);
-            if (id.isEmpty()) {
-                throw new ValidationException("Invalid $id value", "$id", "[empty string]");
-            }
-            final URI uri = new URI(id);
-            loaderBuilder.registerSchemaByURI(uri, schema);
-        } catch (URISyntaxException e) {
-            throw new ValidationException("Invalid $id value", "$id", e);
-        } catch (JSONException e) {
-            // $id is missing or not a string
-            throw new ValidationException("Invalid $id value", "$id", e);
         }
     }
 }
