+++ src/test/java/software/amazon/cloudformation/resource/ValidatorTest.java
@@ -14,38 +14,31 @@
 */
 package software.amazon.cloudformation.resource;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.assertj.core.api.Assertions.catchThrowableOfType;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 import java.util.stream.Stream;
 
-import org.everit.json.schema.loader.SchemaLoader;
 import org.json.JSONObject;
 import org.json.JSONTokener;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import software.amazon.cloudformation.resource.exceptions.ValidationException;
 
 public class ValidatorTest {
-    private static final String RESOURCE_DEFINITION_SCHEMA_PATH = "/schema/provider.definition.schema.v1.json";
     private static final String TEST_SCHEMA_PATH = "/test-schema.json";
-    private static final String TEST_VALUE_SCHEMA_PATH = "/scrubbed-values-schema.json";
     private static final String SCHEMA_WITH_HANDLERS_PATH = "/valid-with-handlers-schema.json";
     private static final String TYPE_NAME_KEY = "typeName";
     private static final String PROPERTIES_KEY = "properties";
+    private static final String TYPE_CONFIGURATION_KEY = "typeConfiguration";
     private static final String DESCRIPTION_KEY = "description";
     private static final String PRIMARY_IDENTIFIER_KEY = "primaryIdentifier";
     private static final String ADDITIONAL_PROPERTIES_KEY = "additionalProperties";
@@ -60,6 +53,10 @@ public class ValidatorTest {
         return baseSchema(new JSONObject().put("property", new JSONObject()));
     }
 
+    private JSONObject baseSchemaWithTypeConfiguration() {
+        return baseSchema(new JSONObject().put("property", new JSONObject()), new JSONObject().put("property", new JSONObject()));
+    }
+
     private JSONObject baseSchema(final JSONObject propertiesValue) {
         return new JSONObject().put(TYPE_NAME_KEY, EXAMPLE_TYPE_NAME).put(DESCRIPTION_KEY, EXAMPLE_DESCRIPTION)
             .put(PROPERTIES_KEY, propertiesValue)
@@ -67,224 +64,18 @@ public class ValidatorTest {
             .put(ADDITIONAL_PROPERTIES_KEY, ADDITIONAL_PROPERTIES_VALUE);
     }
 
+    private JSONObject baseSchema(final JSONObject propertiesValue, final JSONObject typeConfigurationValue) {
+        return new JSONObject().put(TYPE_NAME_KEY, EXAMPLE_TYPE_NAME).put(DESCRIPTION_KEY, EXAMPLE_DESCRIPTION)
+            .put(PROPERTIES_KEY, propertiesValue)
+            .put(PRIMARY_IDENTIFIER_KEY, Collections.singletonList(EXAMPLE_PRIMARY_IDENTIFIER))
+            .put(ADDITIONAL_PROPERTIES_KEY, ADDITIONAL_PROPERTIES_VALUE)
+            .put(TYPE_CONFIGURATION_KEY, new JSONObject().put(ADDITIONAL_PROPERTIES_KEY, ADDITIONAL_PROPERTIES_VALUE)
+                .put(PROPERTIES_KEY, typeConfigurationValue));
+    }
+
     @BeforeEach
     public void setUp() {
         validator = new Validator();
-    }
-
-    @Test
-    public void validateObject_validObject_shouldNotThrow() {
-        final JSONObject object = new JSONObject().put("propertyA", "abc").put("propertyB", Arrays.asList(1, 2, 3));
-
-        validator.validateObject(object, new JSONObject(new JSONTokener(this.getClass().getResourceAsStream(TEST_SCHEMA_PATH))));
-    }
-
-    @Test
-    public void validateObject_invalidObjectMissingRequiredProperties_shouldThrow() {
-        final String propVal = "abc";
-        final JSONObject object = new JSONObject().put("propertyA", propVal);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object,
-                new JSONObject(new JSONTokener(this.getClass().getResourceAsStream(TEST_SCHEMA_PATH)))),
-            ValidationException.class);
-
-        assertThat(e).hasNoCause().hasMessageContaining("propertyB").hasMessageNotContaining(propVal);
-        assertThat(e.getCausingExceptions()).isEmpty();
-        assertThat(e.getSchemaPointer()).isEqualTo("#");
-        assertThat(e.getKeyword()).isEqualTo("required");
-    }
-
-    @Test
-    public void validateObject_invalidObjectAdditionalProperties_shouldThrow() {
-        final String propValue = "notpartofschema";
-        final JSONObject object = new JSONObject().put("propertyA", "abc").put("propertyB", Arrays.asList(1, 2, 3))
-            .put("propertyX", propValue);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object,
-                new JSONObject(new JSONTokener(this.getClass().getResourceAsStream(TEST_SCHEMA_PATH)))),
-            ValidationException.class);
-
-        assertThat(e).hasNoCause().hasMessageContaining("propertyX");
-        assertThat(e).hasMessageNotContaining(propValue);
-        assertThat(e.getCausingExceptions()).isEmpty();
-        assertThat(e.getSchemaPointer()).isEqualTo("#");
-        assertThat(e.getKeyword()).isEqualTo("additionalProperties");
-    }
-
-    @Test
-    public void validateObject_invalidType_messageShouldNotContainValue() {
-        final JSONObject object = new JSONObject().put("BooleanProperty", "true");
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object,
-                new JSONObject(new JSONTokener(this.getClass().getResourceAsStream(TEST_VALUE_SCHEMA_PATH)))),
-            ValidationException.class);
-
-        assertThat(e).hasMessageNotContaining("true");
-        assertThat(e.getCausingExceptions()).isEmpty();
-        assertThat(e.getKeyword()).isEqualTo("type");
-    }
-
-    @ParameterizedTest
-    @CsvSource({ "WaaaaaaaayTooLong,maxLength", "TooShort,minLength", "NoPatternMatch,pattern" })
-    public void validateObject_invalidStringValue_messageShouldNotContainValue(final String value, final String keyword) {
-        final JSONObject object = new JSONObject().put("StringProperty", value);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object, loadJSON(TEST_VALUE_SCHEMA_PATH)),
-            ValidationException.class);
-
-        assertThat(e.getSchemaPointer()).isEqualTo("#/StringProperty");
-        assertThat(e.getKeyword()).isEqualTo(keyword);
-        assertThat(e.getMessage()).doesNotContain(value);
-        assertThat(e.getCausingExceptions()).isEmpty();
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = { "enum", "const" })
-    public void validateObject_invalidEnumValue_messageShouldNotContainValue(final String keyword) {
-        final String propName = keyword + "Property";
-        final String propVal = "NotPartOfEnum";
-        final JSONObject object = new JSONObject().put(propName, propVal);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object,
-                new JSONObject(new JSONTokener(this.getClass().getResourceAsStream(TEST_VALUE_SCHEMA_PATH)))),
-            ValidationException.class);
-
-        final String pointer = "#/" + propName;
-        assertThat(e.getSchemaPointer()).isEqualTo(pointer);
-        assertThat(e.getKeyword()).isEqualTo("allOf");
-        assertThat(e).hasMessageNotContaining(propVal);
-        assertThat(e.getCausingExceptions()).hasSize(1);
-
-        final ValidationException enumEx = e.getCausingExceptions().get(0);
-        assertThat(enumEx.getSchemaPointer()).isEqualTo(pointer);
-        assertThat(enumEx).hasMessageNotContaining(propVal);
-        assertThat(enumEx.getKeyword()).isEqualTo(keyword);
-        assertThat(enumEx.getCausingExceptions()).isEmpty();
-    }
-
-    @ParameterizedTest
-    @CsvSource({ "test-test,uniqueItems", "test,minItems", "test-test2-test3,maxItems", "Y-X,contains" })
-    public void validateObject_invalidArrayValue_messageShouldNotContainValue(final String listAsString, final String keyword) {
-        final List<String> values = Arrays.asList(listAsString.split("-"));
-        final JSONObject object = new JSONObject().put("ArrayProperty", values);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object, loadJSON(TEST_VALUE_SCHEMA_PATH)),
-            ValidationException.class);
-
-        assertThat(e.getKeyword()).isEqualTo(keyword);
-        assertThat(e.getSchemaPointer()).isEqualTo("#/ArrayProperty");
-        assertThat(e.getCausingExceptions()).isEmpty();
-
-        values.forEach(v -> assertThat(e).hasMessageNotContaining(v));
-    }
-
-    @ParameterizedTest
-    @CsvSource({ "5,IntProperty,minimum", "300,IntProperty,maximum", "23,IntProperty,multipleOf",
-        "5,NumberProperty,exclusiveMinimum", "300,NumberProperty,exclusiveMaximum" })
-    public void validateObject_invalidNumValue_messageShouldNotContainValue(final String numAsString,
-                                                                            final String propName,
-                                                                            final String keyword) {
-        final JSONObject object = new JSONObject().put(propName, Integer.valueOf(numAsString));
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object, loadJSON(TEST_VALUE_SCHEMA_PATH)),
-            ValidationException.class);
-
-        assertThat(e.getKeyword()).isEqualTo(keyword);
-        assertThat(e.getSchemaPointer()).isEqualTo("#/" + propName);
-        assertThat(e).hasMessageNotContaining(numAsString);
-        assertThat(e.getCausingExceptions()).isEmpty();
-    }
-
-    @ParameterizedTest
-    @CsvSource({ "test,minProperties", "test-test1-test2,maxProperties", "test-dep,dependencies" })
-    public void validateObject_invalidSubObject_messageShouldNotContainValue(final String keysAsString, final String keyword) {
-        final String val = "testValue";
-        final JSONObject subSchema = new JSONObject();
-
-        final List<String> keys = Arrays.asList(keysAsString.split("-"));
-        keys.forEach(k -> subSchema.put(k, val));
-
-        final JSONObject object = new JSONObject().put("ObjectProperty", subSchema);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object,
-                new JSONObject(new JSONTokener(this.getClass().getResourceAsStream(TEST_VALUE_SCHEMA_PATH)))),
-            ValidationException.class);
-
-        assertThat(e.getKeyword()).isEqualTo(keyword);
-        assertThat(e.getSchemaPointer()).isEqualTo("#/ObjectProperty");
-        assertThat(e).hasMessageNotContaining(val);
-        assertThat(e.getCausingExceptions()).isEmpty();
-    }
-
-    @ParameterizedTest
-    @CsvSource({ "test,minProperties", "test-test1-test2,maxProperties", "test-dep,dependencies" })
-    public void validateObject_invalidPatternProperties_messageShouldNotContainValue(final String keysAsString,
-                                                                                     final String keyword) {
-        final String val = "Value";
-        final JSONObject object = new JSONObject().put("MapProperty", new JSONObject().put("def", "val"));
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object, loadJSON(TEST_VALUE_SCHEMA_PATH)),
-            ValidationException.class);
-
-        assertThat(e.getSchemaPointer()).isEqualTo("#/MapProperty");
-        assertThat(e.getKeyword()).isEqualTo("additionalProperties");
-        assertThat(e).hasMessageNotContaining(val);
-        assertThat(e.getCausingExceptions()).isEmpty();
-
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = { "anyOf", "allOf", "oneOf" })
-    public void validateObject_invalidCombiner_messageShouldNotContainValue(final String keyword) {
-        final String propName = keyword + "Property";
-        final String propVal = "NotAnInteger";
-        final JSONObject object = new JSONObject().put(propName, propVal);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object, loadJSON(TEST_VALUE_SCHEMA_PATH)),
-            ValidationException.class);
-
-        final String pointer = "#/" + propName;
-        assertThat(e.getSchemaPointer()).isEqualTo(pointer);
-        assertThat(e.getKeyword()).isEqualTo(keyword);
-        assertThat(e).hasMessageNotContaining(propVal);
-        assertThat(e.getCausingExceptions()).hasSize(1);
-
-        final ValidationException enumEx = e.getCausingExceptions().get(0);
-        assertThat(enumEx.getSchemaPointer()).isEqualTo(pointer);
-        assertThat(enumEx).hasMessageNotContaining(propVal);
-        assertThat(enumEx.getKeyword()).isEqualTo("type");
-        assertThat(enumEx.getCausingExceptions()).isEmpty();
-    }
-
-    @Test
-    public void validateObject_invalidObjectMultiple_messageShouldNotContainValue() {
-        final String propValue = "notpartofschema";
-        final JSONObject object = new JSONObject().put("propertyA", 123).put("propertyB", Arrays.asList(1, 2, 3))
-            .put("propertyX", propValue).put("propertyY", propValue);
-
-        final ValidationException e = catchThrowableOfType(
-            () -> validator.validateObject(object, loadJSON(TEST_SCHEMA_PATH)),
-            ValidationException.class);
-
-        assertThat(e.getCausingExceptions()).hasSize(3);
-        assertThat(e).hasMessage("#: 3 schema violations found");
-        assertThat(e.getSchemaPointer()).isEqualTo("#");
-        assertThat(e.getKeyword()).isNull();
-
-        e.getCausingExceptions().forEach(ce -> {
-            assertThat(ce).hasMessageNotContaining(propValue);
-            assertThat(ce.getCausingExceptions()).isEmpty();
-        });
     }
 
     @Test
@@ -498,30 +289,53 @@ public class ValidatorTest {
         validator.validateResourceDefinition(example);
     }
 
-    /**
-     * trivial coverage test: cannot cache a schema if it has an invalid $id
-     */
     @ParameterizedTest
-    @ValueSource(strings = { ":invalid/uri", "" })
-    public void registerMetaSchema_invalidRelativeRef_shouldThrow(String uri) {
+    @ValueSource(strings = { PROPERTIES_KEY, ADDITIONAL_PROPERTIES_KEY })
+    public void validateDefinition_withTypeConfiguration_requiredKeyMissing_shouldThrow(final String requiredKey) {
+        final JSONObject schema = baseSchemaWithTypeConfiguration();
+        JSONObject typeConfiguration = schema.getJSONObject(TYPE_CONFIGURATION_KEY);
+        typeConfiguration.remove(requiredKey);
+        schema.put(TYPE_CONFIGURATION_KEY, typeConfiguration);
 
-        JSONObject badSchema = loadJSON(RESOURCE_DEFINITION_SCHEMA_PATH);
-        badSchema.put("$id", uri);
-        assertThatExceptionOfType(ValidationException.class).isThrownBy(() -> {
-            validator.registerMetaSchema(SchemaLoader.builder(), badSchema);
-        });
+        assertThatExceptionOfType(ValidationException.class).isThrownBy(
+            () -> validator.validateResourceDefinition(schema))
+            .withNoCause().withMessageContaining(requiredKey);
     }
 
-    /**
-     * trivial coverage test: cannot cache a schema if it has no $id
-     */
+    @ParameterizedTest
+    @ValueSource(strings = { "CreateOnlyProperties", "PrimaryIdentifier", "Random" })
+    public void validateDefinition_withTypeConfiguration_extraneousKeys_shouldThrow(final String extraneousKey) {
+        final JSONObject schema = baseSchemaWithTypeConfiguration();
+        JSONObject typeConfiguration = schema.getJSONObject(TYPE_CONFIGURATION_KEY);
+        typeConfiguration.put(extraneousKey, new JSONObject());
+        schema.put(TYPE_CONFIGURATION_KEY, typeConfiguration);
+
+        assertThatExceptionOfType(ValidationException.class).isThrownBy(
+            () -> validator.validateResourceDefinition(schema))
+            .withNoCause().withMessageContaining(extraneousKey);
+    }
+
     @Test
-    public void registerMetaSchema_nullId_shouldThrow() {
-        JSONObject badSchema = loadJSON(RESOURCE_DEFINITION_SCHEMA_PATH);
-        badSchema.remove("$id");
-        assertThatExceptionOfType(ValidationException.class).isThrownBy(() -> {
-            validator.registerMetaSchema(SchemaLoader.builder(), badSchema);
-        });
+    public void validateDefinition_withTypeConfiguration_invalidDefinitionNoProperties_shouldThrow() {
+        final JSONObject definition = baseSchema(new JSONObject().put("test", new JSONObject().put("type", "string")),
+            new JSONObject());
+
+        assertThatExceptionOfType(ValidationException.class).isThrownBy(
+            () -> validator.validateResourceDefinition(definition))
+            .withNoCause().withMessageContaining("#/typeConfiguration/properties: minimum size: [1], found: [0]");
+    }
+
+    @Test
+    public void validateDefinition_idKeyword_shouldNotBeAllowed() {
+        final JSONObject schema = baseSchemaWithTypeConfiguration();
+        final JSONObject typeConfiguration = schema.getJSONObject(TYPE_CONFIGURATION_KEY);
+        typeConfiguration.put("$id",
+            "https://schema.cloudformation.us-east-1.amazonaws.com/aws-ec2-instance.json#");
+        schema.put(TYPE_CONFIGURATION_KEY, typeConfiguration);
+
+        assertThatExceptionOfType(ValidationException.class).isThrownBy(
+            () -> validator.validateResourceDefinition(schema))
+            .withNoCause().withMessageContaining("#/typeConfiguration: extraneous key [$id] is not permitted");
     }
 
     private static Stream<Arguments> generateValidReplacementStrategies() {
