+++ src/test/resources/minimal-schema-with-typeconfiguration.json
@@ -1,0 +1,24 @@
+{
+    "typeName": "AWS::Test::TestModel",
+    "description": "A test schema for unit tests.",
+    "properties": {
+        "propertyA": {
+            "type": "boolean"
+        }
+    },
+    "primaryIdentifier": [
+        "/properties/PropertyA"
+    ],
+    "additionalProperties": false,
+    "typeConfiguration": {
+        "additionalProperties": false,
+        "properties": {
+            "Configuration": {
+                "type": "string"
+            },
+            "APIKey": {
+                "type": "string"
+            }
+        }
+    }
+}
